Starting the router
Started constructing the router for #1

we have a list of special namespaces to take care of the admin section. Can be extended if we need other sections just by adding the section to the private $section array.

we can call methods with camelcase like so :
blogoc/post/add-post
will convert to
Post->addPost()

all that is passed in the url is converted to lowercase to respect the strict naming convention without bothering the visitor.

Still need to take care of the paramters to be passed to the methods.