cleanup
Cleaned up some old tests and refactoring the new method
need to find a better way rather than 1000 includes.
Prehaps a use in the core Controller method ?
not sure if that is OK since the core will be in relation with the traits. and the traits will use the Models.
Method reflection needed ...