moved the menu implementation. No more Traits
Ok, Brain fart. Just put all the stuff in models. A lot simpler.
Then call some "if var is defined" in twig and include the sub template.

A lot more flexable. the controller "Controls" what will be displayed and the view just displays what is needed.

The only annoying thing is that we have to create and call the models each time. Could make them static to avoid the "new" object creation call but then i would be mixing static and non static calls from the model namespace. feals dirty.

So just need to clean up (probalby delete this branch and redo propper)