finished the generic get table name ad adding some generic querys
The get table name function is finished.
If no table name is passed, we check for a table with the model name.
If none exist, we check for a view with v_modelName.

We can also pass in a table or view name. The program will still do the check.

Also added a few generic selects :
getresultSet -> gets the entire table
getRowById -> pass in an ID and it returns the line
getRowByColumn -> pass in a column name and a value and it returns the coresponding line.

The last function probably needs some sanitizing to avoid SQL injections on the columnname as PDO doesn't alow prepared statments with column or table names

Also need to docu the functions