Setting up the security for ajax calls
advancing on #10
added CSRF class to handle security. It creates the csrf_token as soon as it is initiated.

what happens is
*The token is generated and passed to the view.
*The view stors the token in a meta in the head
*the JS ajax query sends that token back to the server
*the server validates that the token is the same as the token in $_SESSION

Moved the Request and Session dependencey classes to a sub folder.
(They should only be called by our container.)

We now have an AjaxController that extends the main Controller. All the ajax controllers should extend this.
The AjaxController starts all the security checks as soon as it's constructed.

The main controller loads the Csrf object on construct to generate the token and adds the token to data['csrf_token'] that will be sent down to the view.