Updating the render functions in controller
We are storing the $data in our mother object.
Just send that rather than having the children fuss with the $this->data.
All that is stored in data will now be passed to the template. That way we are sure the CSRF token is sent.

Also need the PDO in the model because we are calling constants.