better login logout test + default controller upgrade
Adding some functions to test the different status (admin, user, visitor). This is just for testing while using the sessions. Moving on to the actual componants of the site to validate the core componants #30

also adding a default index to the core controller that will redirect to index in prod or throw an error if in dev for easier debugging

And finaly adding a redirect to the response dependency