Added some login helpers for easier demo + corrected a few bugs
Added some login helpers to the connection window just for testing.

Also changeg the toastr positionning since it coverd up the top bars.

Corrected an error if we passed a redirect with a forward slash

Automaticly directed to the admin panel on connection

Hocked up the disconnect button in the admin pannel (none of the others are set yet)

Also added a link to the admin pannel to return to the site