adding a check if in dev mode and multiple modules declared
We could have multiple modules declared whith app or namespace modules over-charging the core and app modules.
We now have an error message to warn if in dev mode.
in prod mode, we should know what we are doing so no message.
And meny if's because of codacy since it hates the 'else'