taking care of table prefixes
Cleaned up the sql query a bit. No need to append the table name.
Ok, for every query we do, we must pass the  table through the model prefixer.

Could also send throught the getTable but that fires a few DB requests to check if the table is present which could slow the process down.