finalising the update config methods
Pushing our set config to the database. this is done in pure php and can be used as a fallback if no javascript active.

some security checks might be needed bet have to be careful with escapting since we do have a url for an image

Also updated the sql to include the type of config. We shall use this for security checks and image uploader

Have to do an ajaxcall to the same model for the js version.