Finished the Ajax calls to theconfig update and integrated on front page
The config reading and updating is functional in pure php and ajax.

I will have to look over it all again with a clear mind to check for security issues

also create a module to read the config since it will be needed in most controllers and the database table pull wasn't usable in the state so doing the conversion in the module as well.

And did a spot of code cleanup