Creating the front posts
Added constants for number of posts and number of front posts; these are the number of posts thet willbe displayed.
Added a model to fetch the posts from the database. Aslo specify if front page articles or not.

Added an excerpt function to the stringfunction trait
added a excerpt word count to the constants
adding all that to the controller and showing on the front page

Updated the DB with test elements